+++ lib/config/bloc_provider/register_reusable_editor_bloc_provider.dart
@@ -7,21 +7,21 @@ import 'package:reusable_editor/reusable_editor.dart';
 
 late TextFieldCubit _textFieldCubit;
 late DateTimeCubit _dateTimeCubit;
-late ImageCrudCubit _imageCrudCubit;
-late SearchFieldCubit _searchFieldCubit;
+late ImageFieldCubit _imageCrudCubit;
 late SwitchCubit _switchCubit;
 late EnumOptionCubit _selectionCubit;
 late ToggleCubit _toggleCubit;
+// late SearchFieldCubit _searchFieldCubit;
 
 initReusableEditorBlocProvider() {
   // ---- Reusable Editor Init ------
   _textFieldCubit = sl<TextFieldCubit>();
   _dateTimeCubit = sl<DateTimeCubit>();
-  _imageCrudCubit = sl<ImageCrudCubit>();
-  _searchFieldCubit = sl<SearchFieldCubit>();
+  _imageCrudCubit = sl<ImageFieldCubit>();
   _switchCubit = sl<SwitchCubit>();
   _selectionCubit = sl<EnumOptionCubit>();
   _toggleCubit = sl<ToggleCubit>();
+  // _searchFieldCubit = sl<SearchFieldCubit>();
 }
 
 disposeReusableEditorBlocProvider() {
@@ -29,19 +29,19 @@ disposeReusableEditorBlocProvider() {
   _textFieldCubit.close();
   _dateTimeCubit.close();
   _imageCrudCubit.close();
-  _searchFieldCubit.close();
   _switchCubit.close();
   _selectionCubit.close();
   _toggleCubit.close();
+  // _searchFieldCubit.close();
 }
 
 List<SingleChildWidget> reusableEditorBlocProviders = [
   // ---- Reusable Editor Bloc Provider ------
   BlocProvider<TextFieldCubit>(create: (_) => _textFieldCubit),
   BlocProvider<DateTimeCubit>(create: (_) => _dateTimeCubit),
-  BlocProvider<ImageCrudCubit>(create: (_) => _imageCrudCubit),
-  BlocProvider<SearchFieldCubit>(create: (_) => _searchFieldCubit),
+  BlocProvider<ImageFieldCubit>(create: (_) => _imageCrudCubit),
   BlocProvider<SwitchCubit>(create: (_) => _switchCubit),
   BlocProvider<EnumOptionCubit>(create: (_) => _selectionCubit),
   BlocProvider<ToggleCubit>(create: (_) => _toggleCubit),
+  // BlocProvider<SearchFieldCubit>(create: (_) => _searchFieldCubit),
 ];