+++ lib/view_models/selection_cubit/enum_option_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:bloc/bloc.dart';
-import 'package:reusable_editor/domain/entities/enum_option_entity.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
+import 'package:reusable_editor/domain/entities/enum_option_entity.dart';
 
 part 'enum_option_state.dart';
 
@@ -11,7 +11,8 @@ class EnumOptionCubit<T extends Enum> extends Cubit<EnumOptionState<T>> {
   // Method to select an option
   void selectOption(EnumOptionEntity<T>? selectedOption) {
     debugPrint(
-        'SelectionCubit | selectOption | type: ${selectedOption?.type?.toString()}');
+      'SelectionCubit | selectOption | type: ${selectedOption?.type?.toString()}',
+    );
     if (selectedOption != null) {
       emit(EnumOptionState(selectedOption: selectedOption));
     }