+++ lib/view_models/text_field_cubit/text_field_cubit.dart
@@ -6,19 +6,22 @@ import 'package:i_validator/i_validator.dart';
 part 'text_field_state.dart';
 
 class TextFieldCubit extends Cubit<TextFieldState> {
-  TextFieldCubit({IValidator? validator})
-    : super(TextFieldState(iValidator: validator ?? RequiredFieldValidator()));
+  final IValidator<String> validator;
 
-  void onChanged(String txt) {
+  TextFieldCubit({IValidator<String>? validator})
+    : validator = validator ?? RequiredFieldValidator(),
+      super(const TextFieldState());
+
+  void onChanged(String? txt) {
     debugPrint('TextFieldCubit | onChanged | txt: $txt');
-    validate(txt);
-    if (state.isValidInput) {
-      emit(state.copyWith(txt: txt));
-    }
+    final error = validator.validate(txt);
+    emit(state.copyWith(txt: txt, txtFieldError: error));
   }
 
+  void clear() => emit(state.clear());
+
   String? validate(String? txt) {
-    final String? error = state.iValidator?.validate(txt ?? '');
+    final error = validator.validate(txt ?? '');
     emit(state.copyWith(txtFieldError: error));
     return error;
   }